sct: split out sct manager crate from chain crate
