+++ crates/sct/Cargo.toml
@@ -1,20 +1,14 @@
 [package]
-name = "penumbra-shielded-pool"
+# TODO: merge with tct crate under a `component` feature flag?
+name = "penumbra-sct"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-component = [
-    "penumbra-component", 
-    "penumbra-storage", 
-    "penumbra-proto/penumbra-storage", 
-    "penumbra-chain/component", 
-    "penumbra-sct/component"
-]
-proving-keys = ["penumbra-proof-params/proving-keys"]
-default = ["std", "component", "proving-keys"]
+component = ["penumbra-component", "penumbra-storage", "penumbra-proto/penumbra-storage", "penumbra-chain/component"]
+default = ["std", "component"]
 std = ["ark-ff/std"]
 docsrs = []
 
@@ -22,12 +16,10 @@ docsrs = []
 # Workspace dependencies
 penumbra-crypto = { path = "../../crypto", default-features = false }
 penumbra-proto = { path = "../../proto", default-features = false }
-penumbra-sct = { path = "../sct", default-features = false }
+penumbra-chain = { path = "../../chain", default-features = false }
 penumbra-component = { path = "../component", optional = true }
 penumbra-storage = { path = "../../storage", optional = true }
-penumbra-chain = { path = "../../chain", default-features = false }
 penumbra-tct = { path = "../../tct" }
-penumbra-proof-params = { path = "../../proof-params", default-features = false }
 
 # Penumbra dependencies
 decaf377-rdsa = { version = "0.5" }
@@ -44,5 +36,7 @@ blake2b_simd = "0.5"
 bytes = "1"
 rand_core = { version = "0.6.3", features = ["getrandom"] }
 rand = "0.8"
+bincode = "1.3.3"
+
