+++ crates/sct/src/lib.rs
@@ -1,0 +1,8 @@
+#![cfg_attr(docsrs, feature(doc_cfg))]
+
+#[cfg_attr(docsrs, doc(cfg(feature = "component")))]
+#[cfg(feature = "component")]
+pub mod component;
+
+pub mod event;
+pub mod state_key;
