+++ pcli/Cargo.toml
@@ -38,6 +38,7 @@ penumbra-tct = { path = "../tct" }
 penumbra-shielded-pool = { path = "../crates/shielded-pool", default-features = false }
 penumbra-ibc = { path = "../crates/ibc", default-features = false }
 penumbra-dao = { path = "../crates/dao", default-features = false }
+penumbra-sct = { path = "../crates/sct", default-features = false }
 # TODO: replace by a penumbra-app
 penumbra-app = { path = "../app" }
