+++ pcli/src/command/query/shielded_pool.rs
@@ -41,13 +41,12 @@ pub enum ShieldedPool {
 impl ShieldedPool {
     pub fn key(&self) -> String {
         use penumbra_chain::state_key as chain_state_key;
+        use penumbra_sct::state_key as sct_state_key;
         use penumbra_shielded_pool::state_key;
         match self {
-            ShieldedPool::Anchor { height } => chain_state_key::anchor_by_height(*height),
-            ShieldedPool::BlockAnchor { height } => {
-                chain_state_key::block_anchor_by_height(*height)
-            }
-            ShieldedPool::EpochAnchor { epoch } => chain_state_key::epoch_anchor_by_index(*epoch),
+            ShieldedPool::Anchor { height } => sct_state_key::anchor_by_height(*height),
+            ShieldedPool::BlockAnchor { height } => sct_state_key::block_anchor_by_height(*height),
+            ShieldedPool::EpochAnchor { epoch } => sct_state_key::epoch_anchor_by_index(*epoch),
             ShieldedPool::CompactBlock { height } => chain_state_key::compact_block(*height),
             ShieldedPool::Commitment { commitment } => state_key::note_source(commitment),
             ShieldedPool::Nullifier { nullifier } => state_key::spent_nullifier_lookup(nullifier),
